refactor: cleanup popover
